+++ packages/international-phone-input/src/Component.responsive.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 
-import { useMatchMedia } from '@alfalab/core-components-mq';
-import { getComponentBreakpoint } from '@alfalab/core-components-shared';
+import { useIsDesktop } from '@alfalab/core-components-mq';
 
 import { InternationalPhoneInputDesktop } from './desktop';
 import { InternationalPhoneInputMobile } from './mobile';
@@ -10,14 +9,24 @@ import { InternationalPhoneInputProps } from './types';
 export const InternationalPhoneInput = React.forwardRef<
     HTMLInputElement,
     InternationalPhoneInputProps
->(({ breakpoint = getComponentBreakpoint(), defaultMatchMediaValue, ...restProps }, ref) => {
-    const [isDesktop] = useMatchMedia(`(min-width: ${breakpoint}px)`, defaultMatchMediaValue);
+>(
+    (
+        {
+            breakpoint,
+            client,
+            defaultMatchMediaValue = client === undefined ? undefined : client === 'desktop',
+            ...restProps
+        },
+        ref,
+    ) => {
+        const isDesktop = useIsDesktop(breakpoint, defaultMatchMediaValue);
 
-    return isDesktop ? (
-        <InternationalPhoneInputDesktop {...restProps} ref={ref} />
-    ) : (
-        <InternationalPhoneInputMobile {...restProps} ref={ref} />
-    );
-});
+        return isDesktop ? (
+            <InternationalPhoneInputDesktop {...restProps} ref={ref} />
+        ) : (
+            <InternationalPhoneInputMobile {...restProps} ref={ref} />
+        );
+    },
+);
 
 InternationalPhoneInput.displayName = 'InternationalPhoneInput';