+++ packages/markdown/src/typings.ts
@@ -43,7 +43,13 @@ export type MarkdownResponsiveProps = Omit<MarkdownProps, 'platform'> & {
     breakpoint?: number;
 
     /**
+     * Версия, которая будет использоваться при серверном рендеринге
+     */
+    client?: 'desktop' | 'mobile';
+
+    /**
      * Значение по-умолчанию для хука useMatchMedia
+     * @deprecated Используйте client
      */
     defaultMatchMediaValue?: boolean | (() => boolean);
 };