+++ packages/select-with-tags/src/Component.responsive.tsx
@@ -3,16 +3,23 @@ import React, { forwardRef } from 'react';
 import { BottomSheet } from '@alfalab/core-components-bottom-sheet';
 import { FormControlDesktop } from '@alfalab/core-components-form-control/desktop';
 import { FormControlMobile } from '@alfalab/core-components-form-control/mobile';
-import { useMatchMedia } from '@alfalab/core-components-mq';
+import { useIsDesktop } from '@alfalab/core-components-mq';
 import { Popover } from '@alfalab/core-components-popover';
-import { getComponentBreakpoint } from '@alfalab/core-components-shared';
 
 import { BaseSelectWithTags } from './components/base-select-with-tags';
 import { SelectWithTagsProps } from './types';
 
 export const SelectWithTags = forwardRef<HTMLInputElement, SelectWithTagsProps>(
-    ({ breakpoint = getComponentBreakpoint(), defaultMatchMediaValue, ...restProps }, ref) => {
-        const [isDesktop] = useMatchMedia(`(min-width: ${breakpoint}px)`, defaultMatchMediaValue);
+    (
+        {
+            breakpoint,
+            client,
+            defaultMatchMediaValue = client === undefined ? undefined : client === 'desktop',
+            ...restProps
+        },
+        ref,
+    ) => {
+        const isDesktop = useIsDesktop(breakpoint, defaultMatchMediaValue);
 
         return (
             <BaseSelectWithTags