+++ packages/shared/src/getComponentBreakpoint/index.ts
@@ -1,1 +1,0 @@
-export { getComponentBreakpoint } from './getComponentBreakpoint';