+++ packages/typography/src/title-responsive/component.tsx
@@ -1,6 +1,6 @@
 import React, { forwardRef } from 'react';
 
-import { useMatchMedia } from '@alfalab/core-components-mq';
+import { useIsDesktop } from '@alfalab/core-components-mq';
 
 import { Title, TitleProps } from '../title';
 import { TitleMobile } from '../title-mobile';
@@ -11,15 +11,18 @@ export type TitleResponsiveProps = TitleProps & {
      * @default 1024
      */
     breakpoint?: number;
+
+    /**
+     * Версия, которая будет использоваться при серверном рендеринге
+     */
+    client?: 'desktop' | 'mobile';
 };
 
 export const TitleResponsive = forwardRef<
     HTMLHeadingElement | HTMLDivElement,
     TitleResponsiveProps
->(({ defaultMatchMediaValue, breakpoint = 1024, ...restProps }, ref) => {
-    const query = `(min-width: ${breakpoint}px)`;
-
-    const [isDesktop] = useMatchMedia(query, defaultMatchMediaValue);
+>(({ breakpoint, client, ...restProps }, ref) => {
+    const isDesktop = useIsDesktop(breakpoint, client === 'desktop');
 
     const Component = isDesktop ? Title : TitleMobile;