+++ packages/form-control/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 12.7.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.14.0
+
 ## 12.7.0
 
 ### Minor Changes