+++ packages/icon-button/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 6.11.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@11.10.1
+
 ## 6.11.0
 
 ### Minor Changes