+++ packages/navigation-bar-private/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @alfalab/core-components-navigation-bar-private
 
+## 0.5.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.14.0
+    -   button@11.10.1
+    -   icon-button@6.11.1
+
 ## 0.5.0
 
 ### Minor Changes