+++ packages/picker-button/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log
 
+## 11.12.4
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.14.0
+    -   button@11.10.1
+    -   select@17.17.2
+    -   popover@6.3.3
+
+## 11.12.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   select@17.17.1
+
 ## 11.12.2
 
 ### Patch Changes